メソッド移動
