+++ tests/test_api.py
@@ -15,6 +15,7 @@ import requests
 from more_itertools import first_true
 
 import annofabapi
+from annofabapi.api import _create_request_body_for_logger, my_backoff
 from annofabapi.dataclass.annotation import AnnotationV2Output, SimpleAnnotation, SingleAnnotationV2
 from annofabapi.dataclass.annotation_specs import AnnotationSpecsV3
 from annofabapi.dataclass.comment import Comment
@@ -29,10 +30,6 @@ from annofabapi.dataclass.task import Task, TaskHistory
 from annofabapi.models import GraphType, ProjectJobType
 from annofabapi.wrapper import TaskFrameKey
 from tests.utils_for_test import WrapperForTest, create_csv_for_task
-
-# webapiにアクセスするテストモジュール
-pytestmark = pytest.mark.access_webapi
-
 
 inifile = configparser.ConfigParser()
 inifile.read("./pytest.ini", "UTF-8")
@@ -54,6 +51,108 @@ wrapper = service.wrapper
 test_wrapper = WrapperForTest(api)
 
 
+class TestMyBackoff:
+    @my_backoff
+    def requestexception_connectionerror_then_true(self, log: list[Any]):
+        if len(log) == 2:
+            return True
+
+        e: Exception
+        if len(log) == 0:
+            e = requests.exceptions.RequestException()
+        elif len(log) == 1:
+            e = ConnectionError()
+        log.append(e)
+        raise e
+
+    def test_assert_retry(self):
+        log: list[Any] = []
+        assert self.requestexception_connectionerror_then_true(log) is True
+        assert 2 == len(log)
+        print(log)
+        assert isinstance(type(log[0]), requests.exceptions.RequestException)
+        assert isinstance(type(log[1]), ConnectionError)
+
+    @my_backoff
+    def chunkedencodingerror_requestsconnectionerror_then_true(self, log: list[Any]):
+        if len(log) == 2:
+            return True
+
+        e: Exception
+        if len(log) == 0:
+            e = requests.exceptions.ChunkedEncodingError()
+            log.append(e)
+            raise e
+        elif len(log) == 1:
+            e = requests.exceptions.ConnectionError()
+            log.append(e)
+            raise e
+
+    def test_assert_retry2(self):
+        log: list[Any] = []
+        assert self.chunkedencodingerror_requestsconnectionerror_then_true(log) is True
+        assert 2 == len(log)
+        print(log)
+        assert isinstance(type(log[0]), requests.exceptions.ChunkedEncodingError)
+        assert isinstance(type(log[1]), requests.exceptions.ConnectionError)
+
+    @my_backoff
+    def httperror_then_true(self, log: list[Any]):
+        if len(log) == 2:
+            return True
+        response = requests.Response()
+        if len(log) == 0:
+            response.status_code = 429
+            e = requests.exceptions.HTTPError(response=response)
+        elif len(log) == 1:
+            response.status_code = 500
+            e = requests.exceptions.HTTPError(response=response)
+        log.append(e)
+        raise e
+
+    def test_assert_retry_with_httperror(self):
+        log: list[Any] = []
+        assert self.httperror_then_true(log) is True
+        assert 2 == len(log)
+        print(log)
+        assert isinstance(type(log[0]), requests.exceptions.HTTPError)
+        assert log[0].response.status_code == 429
+        assert isinstance(type(log[1]), requests.exceptions.HTTPError)
+        assert log[1].response.status_code == 500
+
+    @my_backoff
+    def httperror_with_400(self, log):
+        if len(log) == 1:
+            return True
+        response = requests.Response()
+        if len(log) == 0:
+            response.status_code = 400
+            e = requests.exceptions.HTTPError(response=response)
+        log.append(e)
+        raise e
+
+    def test_assert_not_retry(self):
+        log: list[Any] = []
+        with pytest.raises(requests.exceptions.HTTPError):
+            self.httperror_with_400(log)
+        assert 1 == len(log)
+
+
+class Test__create_request_body_for_logger:
+    def test_data_dict(self):
+        actual = _create_request_body_for_logger({"foo": "1", "password": "x", "new_password": "y", "old_password": "z"})
+        assert actual == {"foo": "1", "password": "***", "new_password": "***", "old_password": "***"}
+
+    def test_data_dict2(self):
+        actual = _create_request_body_for_logger({"foo": "1"})
+        assert actual == {"foo": "1"}
+
+    def test_data_list(self):
+        actual = _create_request_body_for_logger([{"foo": "1"}])
+        assert actual == [{"foo": "1"}]
+
+
+@pytest.mark.access_webapi
 class TestAnnotation:
     input_data_id: str
 
@@ -111,6 +210,7 @@ class TestAnnotation:
         )
 
 
+@pytest.mark.access_webapi
 class TestAnnotationSpecs:
     def test_get_annotation_specs(self):
         annotation_spec, _ = api.get_annotation_specs(project_id, query_params={"v": "3"})
@@ -140,6 +240,7 @@ class TestAnnotationSpecs:
         assert type(result) is annofabapi.wrapper.AnnotationSpecsRelation
 
 
+@pytest.mark.access_webapi
 class TestComment:
     task: dict[str, Any]
 
@@ -203,6 +304,7 @@ class TestComment:
         wrapper.change_task_status_to_break(project_id, task_id)
 
 
+@pytest.mark.access_webapi
 class TestInputData:
     input_data_id: str
 
@@ -236,6 +338,7 @@ class TestInputData:
         assert type(api.batch_update_inputs(project_id, request_body=request_body)[0]) == list  # noqa: E721
 
 
+@pytest.mark.access_webapi
 class TestInstruction:
     def test_wrapper_get_latest_instruction(self):
         assert type(wrapper.get_latest_instruction(project_id)) == dict  # noqa: E721
@@ -265,6 +368,7 @@ class TestInstruction:
         assert type(api.put_instruction(project_id, request_body=put_request_body)[0]) == dict  # noqa: E721
 
 
+@pytest.mark.access_webapi
 class TestJob:
     @pytest.mark.submitting_job
     def test_scenario(self):
@@ -296,6 +400,7 @@ class TestJob:
         assert first_true(job_list, pred=lambda e: e["job_id"] == job_id) is None
 
 
+@pytest.mark.access_webapi
 class TestLogin:
     def test_login(self):
         # Exceptionをスローしないことの確認
@@ -304,6 +409,7 @@ class TestLogin:
         api.logout()
 
 
+@pytest.mark.access_webapi
 class TestMy:
     def test_get_my_account(self):
         my_account, _ = api.get_my_account()
@@ -326,6 +432,7 @@ class TestMy:
         assert type(my_member_in_project) == dict  # noqa: E721
 
 
+@pytest.mark.access_webapi
 class TestOrganization:
     organization_name: str
 
@@ -346,6 +453,7 @@ class TestOrganization:
         assert len(wrapper.get_all_projects_of_organization(self.organization_name)) > 0
 
 
+@pytest.mark.access_webapi
 class TestOrganizationMember:
     organization_name: str
 
@@ -369,6 +477,7 @@ class TestOrganizationMember:
         api.update_organization_member_role(self.organization_name, api.login_user_id, request_body=request_body)
 
 
+@pytest.mark.access_webapi
 class TestProject:
     def test_get_project(self):
         dict_project, _ = api.get_project(project_id)
@@ -399,6 +508,7 @@ class TestProject:
         assert wrapper.download_project_comments_url(project_id, f"{out_dir}/comments.json").startswith("https://")
 
 
+@pytest.mark.access_webapi
 class TestProjectMember:
     def test_get_project_member(self):
         my_member = api.get_project_member(project_id, api.login_user_id)[0]
@@ -410,6 +520,7 @@ class TestProjectMember:
         ProjectMember.schema().load(member_list, many=True)
 
 
+@pytest.mark.access_webapi
 class TestStatistics:
     def test_wrapper_get_account_daily_statistics(self):
         actual = wrapper.get_account_daily_statistics(project_id, from_date="2021-04-01", to_date="2021-06-30")
@@ -517,6 +628,7 @@ class TestStatistics:
         assert type(content) == list  # noqa: E721
 
 
+@pytest.mark.access_webapi
 class Testsupplementary:
     input_data_id: str
 
@@ -546,6 +658,7 @@ class Testsupplementary:
         assert len([e for e in supplementary_data_list2 if e["supplementary_data_id"] == supplementary_data_id]) == 0
 
 
+@pytest.mark.access_webapi
 class TestTask:
     input_data_id: str
 
@@ -612,6 +725,7 @@ class TestTask:
         assert type(content) == dict  # noqa: E721
 
 
+@pytest.mark.access_webapi
 class TestWebhook:
     def test_scenario(self):
         """
@@ -647,6 +761,7 @@ class TestWebhook:
         assert first_true(webhook_list, pred=lambda e: e["webhook_id"] == test_webhook_id) is None
 
 
+@pytest.mark.access_webapi
 class TestGetObjOrNone:
     """
     wrapper.get_xxx_or_none メソッドの確認
@@ -712,6 +827,7 @@ class TestGetObjOrNone:
         assert supplementary_data_list is None
 
 
+@pytest.mark.access_webapi
 class TestProtectedMethod:
     input_data_id: str
 
@@ -737,6 +853,7 @@ class TestProtectedMethod:
             api._request_get_with_cookie(project_id, url)
 
 
+@pytest.mark.access_webapi
 class TestProperty:
     def test_account_id(self):
         account_id = api.account_id
